fix: 파일 네이밍 컨벤션에 맞춰 component 파일들 수정 -> PascalCase로 통일 #48
